Adapt axisymmetric mappers from 0°-360° towards 0°-self chosen angle (eg: wedge angle in axisymmtric case in OpenFOAM). Unit tests are adapted and tests have been carried out on different example cases with reliable results.
