+++ tests/mappers/test_axisymmetric_2d_to_3d.py
@@ -7,15 +7,16 @@ import matplotlib.pyplot as plt
 from matplotlib import cm
 
 
-class TestMapperAxisymmetric2DTo3D(unittest.TestCase):
-    gui = False
+class TestMapperAxisymmetric2DTo3DMod(unittest.TestCase):
+    gui = True
 
     def setUp(self):
-        self.parameters = {'type': 'mappers.axisymmetric_2d_to_3d',
+        self.parameters = {'type': 'mappers.axisymmetric_2d_to_3d_mod',
                            'settings':
                                {'direction_axial': 'x',
                                 'direction_radial': 'y',
-                                'n_tangential': 13}
+                                'angle':360,
+                                'n_tangential': 10}
                            }
 
     def test_instantiation(self):
@@ -35,6 +36,7 @@ class TestMapperAxisymmetric2DTo3D(unittest.TestCase):
     def test_initialize(self):
         mp_name_in = 'wall_in'
         mp_name_out = 'wall_out'
+        self.angle = self.parameters['settings']['angle']
 
         # create model_part_in
         n_in = 10
@@ -45,7 +47,10 @@ class TestMapperAxisymmetric2DTo3D(unittest.TestCase):
         model.create_model_part(mp_name_in, x_in, y_in, z_in, np.arange(n_in))
 
         # create reference geometry for 3D model_part_out
+
         n_t = self.parameters['settings']['n_tangential']
+        if n_t%2 ==1:
+            n_t += 1
         n_out_ref = n_in * n_t
         x_out_ref = np.zeros(n_out_ref)
         y_out_ref = np.zeros(n_out_ref)
@@ -55,7 +60,11 @@ class TestMapperAxisymmetric2DTo3D(unittest.TestCase):
             for i_from in range(n_in):
                 start = i_t * n_in
                 end = (i_t + 1) * n_in
-                theta = i_t * 2 * np.pi / n_t
+                if i_t % 2 == 1:
+                    theta = (i_t) * ((np.radians(self.angle) * (-1) ** (i_t)) / (2 * n_t))
+                else:
+                    theta = (i_t + 1) * ((np.radians(self.angle) * (-1) ** (i_t)) / (2 * n_t))
+
                 x_out_ref[start:end] = x_in
                 y_out_ref[start:end] = np.cos(theta) * y_in
                 z_out_ref[start:end] = np.sin(theta) * y_in
@@ -93,7 +102,7 @@ class TestMapperAxisymmetric2DTo3D(unittest.TestCase):
         var_s = 'pressure'
         var_v = 'displacement'
 
-        n_from = 7
+        n_from = 10
         tmp = np.linspace(0, 5, n_from)
         x_from, y_from, z_from = tmp, 1. + 0.2 * np.sin(2 * np.pi / 5 * tmp), np.zeros_like(tmp)
         v_s_from = fun_s(x_from).reshape(-1, 1)
@@ -154,6 +163,8 @@ class TestMapperAxisymmetric2DTo3D(unittest.TestCase):
                 ax.set_zlabel('z')
 
             plt.get_current_fig_manager().window.showMaximized()
+            # plt.xlim(0, 6)
+            # plt.ylim(0.7, 1.5)
             plt.show()
             plt.close()
