+++ tests/mappers/test_axisymmetric_3d_to_2d.py
@@ -7,20 +7,22 @@ import matplotlib.pyplot as plt
 from matplotlib import cm
 
 
-class TestMapperAxisymmetric3DTo2D(unittest.TestCase):
-    gui = False
+class TestMapperAxisymmetric3DTo2DMod(unittest.TestCase):
+    gui = True
 
     def setUp(self):
-        self.parameters = {'type': 'mappers.axisymmetric_3d_to_2d',
+        self.parameters = {'type': 'mappers.axisymmetric_3d_to_2d_mod',
                            'settings':
                                {'direction_axial': 'x',
                                 'direction_radial': 'y',
-                                'n_tangential': 13}
+                                'angle':180,
+                                'n_tangential': 3}
                            }
 
     def test_initialize(self):
         mp_name_in = 'wall_in'
         mp_name_out = 'wall_out'
+        self.angle = self.parameters['settings']['angle']
 
         # create model_part_in
         n_in = 10
@@ -31,7 +33,10 @@ class TestMapperAxisymmetric3DTo2D(unittest.TestCase):
         model.create_model_part(mp_name_in, x_in, y_in, z_in, np.arange(n_in))
 
         # create reference geometry for 3D model_part_out
+
         n_t = self.parameters['settings']['n_tangential']
+        if n_t%2 ==1:
+            n_t += 1
         n_out_ref = n_in * n_t
         x_out_ref = np.zeros(n_out_ref)
         y_out_ref = np.zeros(n_out_ref)
@@ -41,7 +46,10 @@ class TestMapperAxisymmetric3DTo2D(unittest.TestCase):
             for i_from in range(n_in):
                 start = i_t * n_in
                 end = (i_t + 1) * n_in
-                theta = i_t * 2 * np.pi / n_t
+                if i_t % 2 == 1:
+                    theta = (i_t) * ((np.radians(self.angle) * (-1) ** (i_t)) / (2 * n_t))
+                else:
+                    theta = (i_t + 1) * ((np.radians(self.angle) * (-1) ** (i_t)) / (2 * n_t))
                 x_out_ref[start:end] = x_in
                 y_out_ref[start:end] = np.cos(theta) * y_in
                 z_out_ref[start:end] = np.sin(theta) * y_in
@@ -79,7 +87,7 @@ class TestMapperAxisymmetric3DTo2D(unittest.TestCase):
         var_s = 'pressure'
         var_v = 'displacement'
 
-        n_to = 7
+        n_to = 10
         tmp = np.linspace(0, 5, n_to)
         x_to, y_to, z_to = tmp, 1. + 0.2 * np.sin(2 * np.pi / 5 * tmp), np.zeros_like(tmp)
